Created ProtoTest.ProtoAST namespace to store all ProtoAST related testcases
Moved relevant testcases to ProtoTest.ProtoAST
