+++ Tests/ProtoTest/ProtoTest.csproj
@@ -134,6 +134,8 @@
     <Compile Include="MultiLangTests\GCTest.cs">
       <SubType>Code</SubType>
     </Compile>
+    <Compile Include="ProtoAST\ProtoASTExecutionTests.cs" />
+    <Compile Include="ProtoAST\RoundTripTests.cs" />
     <Compile Include="TD\Associative\ReplicationGuide.cs" />
     <Compile Include="TD\DocTests\UserManualTests.cs" />
     <Compile Include="TD\Imperative\Imperative.cs" />
